ArchvieFiles: Refactor
